refactored GameOver view to a component
